test: share settings component fixtures
