+++ tests/helpers/MockSettingComponents.ts
@@ -1,0 +1,128 @@
+export interface MockInputElement {
+  rows: number;
+  cols: number;
+  value: string;
+  style: Record<string, string>;
+  dataset: Record<string, string>;
+  addClass: jest.Mock;
+  toggleClass: jest.Mock;
+  addEventListener: jest.Mock;
+}
+
+export interface MockTextComponent {
+  value: string;
+  placeholder: string;
+  onChangeCallback: ((value: string) => Promise<void> | void) | null;
+  setPlaceholder: jest.MockedFunction<(value: string) => MockTextComponent>;
+  setValue: jest.MockedFunction<(value: string) => MockTextComponent>;
+  onChange: jest.MockedFunction<(callback: (value: string) => Promise<void> | void) => MockTextComponent>;
+  inputEl: MockInputElement;
+}
+
+export interface MockTextAreaComponent extends MockTextComponent {
+  trigger: (event: string) => Promise<void>;
+}
+
+export interface MockDropdownComponent {
+  value: string;
+  options: Array<{ value: string; label: string }>;
+  onChangeCallback: ((value: string) => Promise<void> | void) | null;
+  addOption: jest.MockedFunction<(value: string, label: string) => MockDropdownComponent>;
+  setValue: jest.MockedFunction<(value: string) => MockDropdownComponent>;
+  onChange: jest.MockedFunction<(callback: (value: string) => Promise<void> | void) => MockDropdownComponent>;
+}
+
+export interface MockToggleComponent {
+  value: boolean;
+  onChangeCallback: ((value: boolean) => Promise<void> | void) | null;
+  setValue: jest.MockedFunction<(value: boolean) => MockToggleComponent>;
+  onChange: jest.MockedFunction<(callback: (value: boolean) => Promise<void> | void) => MockToggleComponent>;
+}
+
+export function createInputElement(): MockInputElement & { _listeners: Map<string, Array<() => void>> } {
+  const listeners = new Map<string, Array<() => void>>();
+  return {
+    rows: 0,
+    cols: 0,
+    value: '',
+    style: {},
+    dataset: {},
+    addClass: jest.fn(),
+    toggleClass: jest.fn(),
+    addEventListener: jest.fn((event: string, handler: () => void) => {
+      const handlers = listeners.get(event) ?? [];
+      handlers.push(handler);
+      listeners.set(event, handlers);
+    }),
+    _listeners: listeners,
+  };
+}
+
+export function createTextComponent(): MockTextComponent {
+  const component = {} as MockTextComponent;
+  component.value = '';
+  component.placeholder = '';
+  component.onChangeCallback = null;
+  component.inputEl = createInputElement();
+  component.setPlaceholder = jest.fn((value: string) => {
+    component.placeholder = value;
+    return component;
+  });
+  component.setValue = jest.fn((value: string) => {
+    component.value = value;
+    component.inputEl.value = value;
+    return component;
+  });
+  component.onChange = jest.fn((callback: (value: string) => Promise<void> | void) => {
+    component.onChangeCallback = callback;
+    return component;
+  });
+  return component;
+}
+
+export function createTextAreaComponent(): MockTextAreaComponent {
+  const component = createTextComponent() as MockTextAreaComponent;
+  component.trigger = async (event: string) => {
+    const listeners = (component.inputEl as ReturnType<typeof createInputElement>)._listeners.get(event) ?? [];
+    for (const handler of listeners) {
+      handler();
+    }
+    await new Promise<void>(resolve => setImmediate(resolve));
+  };
+  return component;
+}
+
+export function createDropdownComponent(): MockDropdownComponent {
+  const component = {} as MockDropdownComponent;
+  component.value = '';
+  component.options = [];
+  component.onChangeCallback = null;
+  component.addOption = jest.fn((value: string, label: string) => {
+    component.options.push({ value, label });
+    return component;
+  });
+  component.setValue = jest.fn((value: string) => {
+    component.value = value;
+    return component;
+  });
+  component.onChange = jest.fn((callback: (value: string) => Promise<void> | void) => {
+    component.onChangeCallback = callback;
+    return component;
+  });
+  return component;
+}
+
+export function createToggleComponent(): MockToggleComponent {
+  const component = {} as MockToggleComponent;
+  component.value = false;
+  component.onChangeCallback = null;
+  component.setValue = jest.fn((value: boolean) => {
+    component.value = value;
+    return component;
+  });
+  component.onChange = jest.fn((callback: (value: boolean) => Promise<void> | void) => {
+    component.onChangeCallback = callback;
+    return component;
+  });
+  return component;
+}
